# Conflicts:
#	Dockerfile
#	README.md
#	docker-compose.yml

+++ releases/maintainer-make-bundle-a.sh
@@ -3,38 +3,27 @@
 # This script is only for Wekan maintainer to
 # convert x64 bundle to arm64 bundle.
 
-cd ~/repos
+if [ $# -ne 1 ]
+  then
+    echo "Syntax with Wekan version number:"
+    echo "  ./maintainer-make-bundle-a.sh 5.10"
+    exit 1
+fi
+
+sudo apt -y install g++ build-essential
+sudo npm -g install node-gyp
 rm -rf bundle
-
+#rm wekan-$1.zip
+#wget https://releases.wekan.team/wekan-$1.zip
 unzip wekan-$1.zip
-
-sudo chown wekan:wekan bundle -R
-sudo apt -y install libcurl4-openssl-dev
-
-sudo rm -f /home/wekan/repos/bundle/programs/server/node_modules/.bin/node-pre-gyp
-sudo rm -f /home/wekan/repos/bundle/programs/server/node_modules/.bin/node-gyp
-sudo rm -rf /home/wekan/repos/bundle/programs/server/npm/node_modules/meteor/lucasantoniassi_accounts-lockout/node_modules/.phantomjs-prebuilt-garbage-*
-sudo rm -rf /home/wekan/repos/bundle/programs/server/node_modules/.bin/*
-sudo rm -rf /home/wekan/repos/bundle/programs/server/node_modules/node-pre-gyp/node_modules/.bin/*
-sudo rm -rf /home/wekan/repos/bundle/programs/server/node_modules/node-gyp/node_modules/.bin/*
-sudo rm -rf /home/wekan/repos/bundle/programs/server/npm/node_modules/meteor/ostrio_files/node_modules/request-libcurl/.node_modules-garbage*
-
 cd bundle/programs/server
 chmod u+w *.json
-npm uninstall fibers node-gyp node-pre-gyp
-npm install node-gyp@5.0.4 node-pre-gyp fibers
-npm install
-npm uninstall fibers node-gyp node-pre-gyp
-npm install node-pre-gyp
-npm install fibers
-npm install node-gyp@5.0.4
-
-cd /home/wekan/repos/bundle
+cd node_modules/fibers
+node build.js
+cd ../../../..
 find . -type d -name '*-garbage*' | xargs rm -rf
 find . -name '*phantom*' | xargs rm -rf
 find . -name '.*.swp' | xargs rm -f
 find . -name '*.swp' | xargs rm -f
-
-cd /home/wekan/repos
-
+cd ..
 zip -r wekan-$1-arm64.zip bundle
