+++ releases/maintainer-make-bundle-o.sh
@@ -1,14 +1,26 @@
-cd /home/ubuntu
+#!/bin/bash
+
+# This script is only for Wekan maintainer to
+# convert x64 bundle to ppc64le bundle.
+
+if [ $# -ne 1 ]
+  then
+    echo "Syntax with Wekan version number:"
+    echo "  ./maintainer-make-bundle-o.sh 5.10"
+    exit 1
+fi
+
+sudo apt -y install g++ build-essential
+sudo npm -g install node-gyp
 rm -rf bundle
+#rm wekan-$1.zip
 #wget https://releases.wekan.team/wekan-$1.zip
 unzip wekan-$1.zip
-cd /home/ubuntu/bundle/programs/server
+cd bundle/programs/server
 chmod u+w *.json
-cd /home/ubuntu/bundle/programs/server/node_modules/fibers
+cd node_modules/fibers
 node build.js
-cd /home/ubuntu
-cp -pR /home/ubuntu/node-fibers/bin/linux-ppc64-72-glibc bundle/programs/server/node_modules/fibers/bin/
-cd bundle
+cd ../../../..
 find . -type d -name '*-garbage*' | xargs rm -rf
 find . -name '*phantom*' | xargs rm -rf
 find . -name '.*.swp' | xargs rm -f
